feat: separate page into upload page and library, add navigation header. fix: correct some flashcard ui features and edit features
